+++ macros/src/mplus/bitmap/from_svgpath/glyph/metrics.rs
@@ -12,6 +12,7 @@ pub struct GlyphMetrics<'a> {
     pub heavy: (Grid<3>, Stroke<'a>),
     pub cross: (Grid<3>, Stroke<'a>),
     pub block: Grid<25>,
+    pub shape: Grid<25>,
 }
 
 impl GlyphMetrics<'_> {
@@ -53,6 +54,8 @@ impl GlyphMetrics<'_> {
         let thin_width_no_snap_grid = Grid::new(width, height, None);
         let zero_width_snap = hint.then_some((0.0, 0.0));
         let zero_width_snap_grid = Grid::new(width, height, zero_width_snap);
+        let half_pixel_width_snap = hint.then_some((0.5, 0.5));
+        let half_pixel_width_snap_grid = Grid::new(width, height, half_pixel_width_snap);
 
         Self {
             halfwidth,
@@ -63,6 +66,7 @@ impl GlyphMetrics<'_> {
             heavy: (thick_width_snap_grid, thick_stroke),
             cross: (thin_width_no_snap_grid, thin_stroke),
             block: zero_width_snap_grid,
+            shape: half_pixel_width_snap_grid,
         }
     }
 }